% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DOR.R
\name{calc_DOR}
\alias{calc_DOR}
\title{Calculate leaf litter Moisture}
\usage{
calc_DOR(Tm_h, vpd_h, rain_h)
}
\arguments{
\item{Tm}{numeric, hourly mean Temperature in degrees}

\item{vpd}{numeric, hourly mean vapour pressure defiecit in hPa}

\item{rain}{numeric, hourly rainfall mean in mm}
}
\description{
Calculate leaf litter Moisture
}
