% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_GER.R
\name{calc_GER}
\alias{calc_GER}
\title{Calculate hourly oospore germination}
\usage{
calc_GER(M_h, T_h)
}
\description{
Calculate the number of germinated oospores in an hour increment using temperature
and moisture. Equation 5 of Rossi (2008)
}
