% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_HT.R
\name{calc_HT}
\alias{calc_HT}
\title{Calculate Hydrothermal Time}
\usage{
calc_HT(Tm_h, M_h)
}
\arguments{
\item{Tm_h}{hourly temperature in degrees}

\item{M_h}{logical, Leaf litter moisture sufficient for maturation}
}
\value{
numeric, Hydrothermal time
}
\description{
Used to determine what the rate at which oospores mature. Described as equation
three in \insertRef{rossi_mechanistic_2008}{viticolR}
}
\examples{
calc_HT(24,1)
}
\references{
\insertAllCited{}
}
