% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PMO.R
\name{calc_PMO}
\alias{calc_PMO}
\title{Calculate physiologically}
\usage{
calc_PMO(hh, MMO, DOR)
}
\arguments{
\item{hh}{hours after midnight on July 1st (Southern Hemisphere), January first
(Northern hemisphere, new years),}

\item{MMO}{Morphologically Mature Oospores at the start of the year}

\item{DOR}{Progress of dormancy converting mophologically mature oospores (MMO)
to Physiologically Mature Oospores (PMO)}
}
\description{
Calculate physiologically
}
\examples{
calc_PMO(123,1,0.002)
}
