% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_DM_PI.R
\name{estimate_DM_PI}
\alias{estimate_DM_PI}
\title{Estimate Downy Mildew Primary Infections}
\usage{
estimate_DM_PI(w, Start = "detect", End = "detect", SOD = 1)
}
\arguments{
\item{w}{hourly weather data.table, class epiphy.weather.}

\item{SOD}{Proportion of seasonal oospores dose. This starts as 1 in the middle
of winter, the first of June for the southern hemisphere or the 1st of January
in the northern hemisphere. Also refered to as 'morphologically mature oospores
(MMO).}
}
\description{
Umbrella function to calculate the number of primary infections on grapevine
leaves from overwintering oospores as described by
\insertRef{rossi_mechanistic_2008}{viticolaR}
}
\references{
\insertAllCited{}
}
