% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_methods.R
\name{geom_line_viticolR}
\alias{geom_line_viticolR}
\title{Plot line viticolR model}
\usage{
geom_line_viticolR(mod, y = "GER", ...)
}
\arguments{
\item{mod}{m_viticolR class object, output of function estimate_DM_PI}

\item{y}{character, An output state variable from the estimate_DM_PI, defaults
to "GER" other options include "PMO", "GEO", "SUZ_h".}

\item{...}{other arguments to be passed to geom_line}
}
\value{
ggplot layer
}
\description{
Plot line viticolR model
}
\examples{
v_mod <- estimate_DM_PI(w = nt_weather,
                        Start = as.POSIXct("2023-07-01"),
                        End = as.POSIXct("2023-08-30"))
ggplot2::ggplot() +
   geom_line_viticolR(v_mod)
}
