% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_methods.R
\name{geom_ribbon_viticolaR}
\alias{geom_ribbon_viticolaR}
\alias{geom_line_viticolaR(mod,}
\alias{y}
\alias{=}
\alias{"GER",}
\alias{...)}
\title{Plot viticolaR model}
\usage{
geom_ribbon_viticolaR(mod, y = "GER", x_subset = "ZooWindow",...)
}
\arguments{
\item{mod}{m_viticola class object, output of function estimate_DM_PI}

\item{y}{character, An output state variable from the estimate_DM_PI, defaults
to "GER" other options include "PMO", "GEO", "SUZ_h".}

\item{x_subset}{a column heading in w_c of the cohort list with a logical class
This will subset the ribbon.}

\item{...}{other arguments to be passed to geom_line}
}
\value{
ggplot
}
\description{
Plot viticolaR model
}
\examples{
v_mod <- estimate_DM_PI(w = nt_weather,
                        Start = as.POSIXct("2023-07-01"),
                        End = as.POSIXct("2023-08-30"))
ggplot2::ggplot() +
   geom_ribbon_viticolaR(v_mod)
}
