% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_methods.R
\name{geom_ribbon_viticolR}
\alias{geom_ribbon_viticolR}
\alias{geom_line_viticolR(mod,}
\alias{y}
\alias{=}
\alias{"GER",}
\alias{...)}
\title{Plot ribbon viticolR model}
\usage{
geom_ribbon_viticolR(mod, y = "GER", x_subset = "ZooWindow",...)
}
\arguments{
\item{mod}{m_viticolR class object, output of function estimate_DM_PI}

\item{y}{character, An output state variable from the estimate_DM_PI, defaults
to "GER" (germinating oospores), other options include "PMO" (Physiological
mature oospores), "GEO" (Germinated oospores), "SUS_h" (survival of sporangia),
"SUZ_h" (Survival of zoospores).}

\item{x_subset}{a column heading in w_c of the cohort list with a logical class
This will subset the ribbon, options include "ZRE_h" (Zoospore release),
"ZDI_h" (Zoospore dispersal hours), "REL" and "ZooWindow" (zoospores on
leaves).}

\item{...}{other arguments to be passed to geom_ribbon}
}
\value{
ggplot
}
\description{
Plot ribbon viticolR model
}
\examples{
v_mod <- estimate_DM_PI(w = nt_weather,
                        Start = as.POSIXct("2023-07-01"),
                        End = as.POSIXct("2023-08-30"))
ggplot2::ggplot() +
   geom_ribbon_viticolR(v_mod)

ggplot2::ggplot() +
   geom_ribbon_viticolR(v_mod,
                         y = "SUZ_h",
                         x_subset = "ZRE_h")
}
