% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PI_dates.R
\name{get_PI_dates}
\alias{get_PI_dates}
\title{Get primary infection dates}
\usage{
get_PI_dates(mod, cohort = "all", release_stage = "all")
}
\arguments{
\item{mod}{model output class 'm_viticola'}

\item{cohort}{integer/s, specify cohort to return the primary infection dates.
'all' can be requested and is by default.}

\item{release_stage}{character vector, specify the release stage/s which you
want returned. "all" by default, or any of the following: "GEO_h",
"SUS_death_h","ZRE_ind", "SUZ_death_ind", "ZDI_ind","ZIN_ind",
"INC_h_lower","INC_h_upper".}
}
\value{
data.table of dates each of the growth stages reach for each cohort
}
\description{
Get primary infection dates
}
\examples{
nt_mod <- estimate_DM_PI(nt_weather)
get_PI_dates(nt_mod)
}
