% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_methods.R
\name{summary.m_viticolR}
\alias{summary.m_viticolR}
\title{Summary of m_viticolR class object}
\usage{
\method{summary}{m_viticolR}(object, ...)
}
\arguments{
\item{object}{m_viticolR class object, output of function estimate_DM_PI}

\item{...}{other arguments to be passed to summary, currently not used.}
}
\value{
a summary of the model run time, weather station, germinated oospore
}
\description{
Summary of m_viticolR class object
}
\examples{
mod1 <- estimate_DM_PI(w = nt_weather,
                       Start = as.Date("2023-07-01"),
                       End = as.Date("2023-08-30"))
summary(mod1)
}
