% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsp_release.R
\name{zsp_release}
\alias{zsp_release}
\title{Zoospore release}
\usage{
zsp_release(WD_h, TWD_h)
}
\arguments{
\item{WD_h}{integer wetness duration in hours}

\item{TWD_h}{numeric, mean temperature during wet duration.}
}
\value{
logical, TRUE or FALSE
}
\description{
Zoospore release
}
