% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SUS.R
\name{calc_SUS}
\alias{calc_SUS}
\title{Calculate sporangia survival}
\usage{
calc_SUS(T_h, RH_h)
}
\arguments{
\item{T_h}{numeric, hourly temperature mean}

\item{RH_h}{numeric, hourly relative humidity mean}
}
\description{
Calculate sporangia survival
}
\examples{
calc_SUS(25,95)
}
