% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_DM_PI.R
\name{estimate_DM_PI}
\alias{estimate_DM_PI}
\title{Estimate Downy Mildew Primary Infections}
\usage{
estimate_DM_PI(w, Start = "detect", End = "detect", SOD = 1)
}
\arguments{
\item{w}{hourly weather data.table, class epiphy.weather.}

\item{Start}{POSIXct, date time in UTC when to start the model running. By default
the function will 'detect' the start of the model. This model is configured
for the southern hemisphere, so the model will start on the first of July of
the first year of weather data input into the model 'w'.}

\item{End}{POSIXct, date time in UTC when to end the model running. By default
the function will 'detect' the end of the model. This model is configured
for the southern hemisphere, so the model will End on the last day of June in
the first season year of the input weather data 'w'.}

\item{SOD}{Proportion of seasonal oospores dose. This starts as 1 in the middle
of winter, the first of June for the southern hemisphere or the 1st of January
in the northern hemisphere. Also referred to as 'morphologically mature oospores
(MMO).}
}
\value{
list, class == m_viticolR, containing the model output and downy mildew
primary infection progress.
\tabular{rl}{
\strong{start_time}: \tab Time in POSIXct format with "UTC" time-zone\cr
\strong{time_hours}: \tab vector of timesteps the model ran \cr
\strong{Hyd_t}: \tab vector of hydrothermal time corresponding to each time_hours \cr
\strong{PMO}: \tab vector of physiological mature oospores at each time step \cr
\strong{cohorts}: \tab total number of oospore cohorts which germinated as a result
of weather conditions meeting the threshold\cr
\strong{w}: \tab weather data.table of class \code{epiphy.weather} used in the model \cr
\strong{cohort_list}: \tab List of model outputs for each cohort \cr
\strong{lat}: \tab Station latitude in decimal degrees \cr
\strong{station}: \tab Unique station identifying name \cr
\strong{YYYY}: \tab Year \cr
\strong{MM}: \tab Month \cr
\strong{DD}: \tab Day \cr
\strong{hh}: \tab Hour \cr
\strong{mm}: \tab Minute \cr}
}
\description{
Umbrella function to calculate the number of primary infections on grapevine
leaves from overwintering oospores as described by
\insertRef{rossi_mechanistic_2008}{viticolR}
}
\examples{
mod1 <- estimate_DM_PI(w = nt_weather,
                       Start = as.Date("2023-07-01"),
                       End = as.Date("2023-08-30"))
}
\references{
\insertAllCited{}
}
